기존의 content 패키지 이름을 admin으로 변경
