+++ app/lib/admin/models/metadata.dart
@@ -25,25 +25,3 @@ class ContentMetadata {
     return 'ContentMetadata{id: $id, title: $title, description: $description, imageUrl: $imageUrl}';
   }
 }
-
-class ContentStreaming {
-  final String platform;
-  final String url;
-
-  const ContentStreaming({
-    required this.platform,
-    required this.url,
-  });
-
-  factory ContentStreaming.fromJson(Map<String, dynamic> json) {
-    return ContentStreaming(
-      platform: json['platform'] as String,
-      url: json['url'] as String,
-    );
-  }
-
-  @override
-  String toString() {
-    return 'ContentStreaming{platform: $platform, url: $url}';
-  }
-}
