+++ app/lib/home/home_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:pancake_app/admin/admin_screen.dart';
 import 'package:pancake_app/api/api.dart' as api;
-import 'package:pancake_app/content/admin_screen.dart';
 import 'package:pancake_app/home/models/catalog.dart';
 import 'package:pancake_app/setting/setting_screen.dart';
 import 'package:url_launcher/url_launcher.dart';
@@ -80,7 +80,7 @@ class _HomeScreenState extends State<HomeScreen> {
             padding: const EdgeInsets.only(right: 20),
             child: ElevatedButton(
               onPressed: _goToAdmin,
-              child: const Text('임시'),
+              child: const Text('어드민'),
             ),
           ),
         ],
